mnp/mnppython/requirements-dev.txt
-r requirements.txt
-e .

pytest
grpcio-tools